requirements.txt
fastapi
uvicorn
python-dotenv
PyGithub
google-generativeai
chromadb
langchain-community  # For text splitting